Remove normalizeThrottleParameters script
I backported everything related to this script to REL1_32 with the hope
to get it ready soon after the release. However, we're now ahead with
REL1_33, and given that this script still has some bugs, the best we can
hope is to have it ready for 1.33 release. Thus, instead of continuing
to backport stuff to REL1_32, remove it altogether.

Bug: T209565
Change-Id: Ie4ff174905bd45bacfeaeca6c72ac25ed0880fc1