Set the CachingParser as the default parser
WMF production was migrated last year. 3rd party wikis have been
receiving deprecation notices for a while. Hence, everything is ready
for this migration. This patch also adds a deprecation notice for
whoever is trying to use the old parser (can only happen if set manually in
the wiki config). As the notice says, the old parser will be removed in
1.36.

Bug: T239990
Change-Id: If44344fc1921ac0460dc72197e1e324dd70c8d23